+++ Packages/com.hung.autotest/Runtime/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Hung.AutoTest.Tests")]
+[assembly: InternalsVisibleTo("Hung.AutoTest.PlayModeTests")]